+++ src/components/issue-list.ts
@@ -1,6 +1,7 @@
 import { getIssueItemTpl } from '~/tpl.js';
 import { findParentAndRenderInnerHtml } from '~/utils/dom';
 import { reduceArray } from '~/utils/array';
+import { ISSUE_LIST_SELECTOR } from '~/constants/selector';
 import type { Issue } from '~/types/issue';
 
 const makeIssueTemplate = ({ issues, templateFunc }: { issues: Issue[], templateFunc: (issue: Issue) => string }): string => {
@@ -18,6 +19,6 @@ const render = ({ parentSelector, issues }: { issues: Issue[], parentSelector: s
   });
 };
 
-export const IssueList = ({ parentSelector, issues }: { issues: Issue[], parentSelector: string }): void => {
+export const IssueList = ({ parentSelector = ISSUE_LIST_SELECTOR, issues }: { issues: Issue[], parentSelector?: string }): void => {
   render({ parentSelector, issues });
 };
